Provide a default neutral template
The old template was silverstrap based, hence if was not working without
silverstrap previously installed. Not a big deal: you would likely
provide your own template anyway, but this not working out-of-the-box
could be a big barrier for newcomers and testers.

The new default template depends only on Fotorama and provides a
skeleton template with a gallery and nothing more. The silverstrap based
template has been moved under silverstrap and will take over this one if
silverstrap itself is installed.